Move Text to separate folder, extract propTypes and defaultProps to separate file and import in Text and EmojiText
